# BLS 12-381 pairing optimization

## Description

This PR aims to improve the pairing for the bls 12-381 curve by using optimized operations


## Type of change


- [ ] Optimization

## Benches
- Actual
`Ate pairing` :  12.169 ms
`Final exponentiation` :  11.499 ms

- New version
`Ate pairing` : 2.0644 ms
`Final exponentiation` : 1.0674 ms